refactor(server): split workspace payload models
